feat: centralize service loading with hook
